+++ pyobs_sbig/__init__.py
@@ -3,9 +3,7 @@ TODO: write doc
 """
 __title__ = 'SBIG camera modules'
 
-from .sbigbasecamera import SbigBaseCamera
-from .sbigdriver import SbigDriver
+from .sbigcamera import SbigBaseCamera
 from .sbigcamera import SbigCamera
-from .sbigtracker import SbigTracker
 from .sbigfiltercamera import SbigFilterCamera
 from .sbig6303e import Sbig6303eCamera
