+++ pyobs_sbig/sbigfiltercamera.py
@@ -26,7 +26,10 @@ class SbigFilterCamera(MotionStatusMixin, SbigCamera, IFilters):
         Args:
             filter_names: List of filter names.
         """
-        SbigCamera.__init__(self, **kwargs, driver_kwargs=dict(filter_wheel=filter_wheel))
+        SbigCamera.__init__(self, **kwargs)
+
+        # filter wheel
+        self.filter_wheel = FilterWheelModel[filter_wheel]
 
         # and filter names
         if filter_names is None:
@@ -52,11 +55,19 @@ class SbigFilterCamera(MotionStatusMixin, SbigCamera, IFilters):
             ValueError: If cannot connect to camera or set filter wheel.
         """
 
+        # set filter wheel model
+        if self.filter_wheel != FilterWheelModel.UNKNOWN:
+            log.info('Initialising filter wheel...')
+            try:
+                self._cam.set_filter_wheel(self.filter_wheel)
+            except ValueError as e:
+                raise ValueError('Could not set filter wheel: %s' % str(e))
+
         # open camera
         await SbigCamera.open(self)
 
         # init status of filter wheel
-        if self._driver.filter_wheel != FilterWheelModel.UNKNOWN:
+        if self.filter_wheel != FilterWheelModel.UNKNOWN:
             await self._change_motion_status(MotionStatus.POSITIONED, interface='IFilters')
 
         # subscribe to events
@@ -82,7 +93,7 @@ class SbigFilterCamera(MotionStatusMixin, SbigCamera, IFilters):
         img = await SbigCamera._expose(self, exposure_time, open_shutter, abort_event)
 
         # add filter to FITS headers
-        if self._driver.filter_wheel != FilterWheelModel.UNKNOWN:
+        if self.filter_wheel != FilterWheelModel.UNKNOWN:
             img.header['FILTER'] = (await self.get_filter(), 'Current filter')
 
         # finished
@@ -100,7 +111,7 @@ class SbigFilterCamera(MotionStatusMixin, SbigCamera, IFilters):
         """
 
         # do we have a filter wheel?
-        if self._driver.filter_wheel == FilterWheelModel.UNKNOWN:
+        if self.filter_wheel == FilterWheelModel.UNKNOWN:
             raise NotImplementedError
 
         # reverse dict and search for name
@@ -109,7 +120,7 @@ class SbigFilterCamera(MotionStatusMixin, SbigCamera, IFilters):
             raise ValueError('Unknown filter: %s' % filter_name)
 
         # there already?
-        position, status = self._driver.camera.get_filter_position_and_status()
+        position, status = self._cam.get_filter_position_and_status()
         if position == filters[filter_name] and status == FilterWheelStatus.IDLE:
             log.info('Filter changed.')
             return
@@ -118,15 +129,15 @@ class SbigFilterCamera(MotionStatusMixin, SbigCamera, IFilters):
         await self._change_motion_status(MotionStatus.SLEWING, interface='IFilters')
 
         # acquire lock
-        with LockWithAbort(self._lock_motion, self._abort_motion):
+        async with LockWithAbort(self._lock_motion, self._abort_motion):
             # set it
             log.info('Changing filter to %s...', filter_name)
-            self._driver.camera.set_filter(filters[filter_name])
+            self._cam.set_filter(filters[filter_name])
 
             # wait for it
             while True:
                 # break, if wheel is idle and filter is set
-                position, status = self._driver.camera.get_filter_position_and_status()
+                position, status = self._cam.get_filter_position_and_status()
                 if position == filters[filter_name] and status == FilterWheelStatus.IDLE:
                     break
 
@@ -154,11 +165,11 @@ class SbigFilterCamera(MotionStatusMixin, SbigCamera, IFilters):
         """
 
         # do we have a filter wheel?
-        if self._driver.filter_wheel == FilterWheelModel.UNKNOWN:
+        if self.filter_wheel == FilterWheelModel.UNKNOWN:
             raise NotImplementedError
 
         try:
-            self._position, _ = self._driver.camera.get_filter_position_and_status()
+            self._position, _ = self._cam.get_filter_position_and_status()
         except ValueError:
             # use existing position
             pass
@@ -175,7 +186,7 @@ class SbigFilterCamera(MotionStatusMixin, SbigCamera, IFilters):
         """
 
         # do we have a filter wheel?
-        if self._driver.filter_wheel == FilterWheelModel.UNKNOWN:
+        if self.filter_wheel == FilterWheelModel.UNKNOWN:
             raise NotImplementedError
 
         # return names
