+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pyobs-sbig"
-version = "0.15.0-alpha.7"
+version = "0.15.0-alpha.8"
 description = "pyobs module for SBIG cameras"
 authors = ["Tim-Oliver Husser <thusser@uni-goettingen.de>"]
 license = "MIT"
